minor bitpack fix and cleanup of cli convert aswell as wip tapresult to parquet
