+++ TapResult.CLI/Converters/Csv.cs
@@ -8,7 +8,7 @@ internal static class Csv
 {
     internal static void Convert(Constants.FileType fileType, Encoder encoder, FileStream input, FileStream output)
     {
-        if (Program.verbose)
+        if (Program.Verbose)
         {
             Console.WriteLine($"Encodings: {string.Join(", ", encoder.EncodingsById.Select(encoding => $"({encoding.Key}: {encoding.Value})"))}");
         }
