+++ TapResult.CLI/Describe.cs
@@ -1,0 +1,6 @@
+namespace TapResult.CLI;
+
+public class Describe
+{
+    
+}
